# Adds IonSystem field to IonContainerLite