@zslayton 2021-05-19
> Side question is there a reasonable unit test we could have that helps replicate or verify this is fixed? I'm guessing not, but just wondering if you had thoughts here.

I spent some time pondering it but didn't have any brilliant ideas. This change shouldn't produce any observable difference in behavior, just in resource usage. We could build a DOM so deeply nested that constructing it would cause a `StackOverflowError`, but that doesn't seem very satisfying.